harm: refactor mod arith macros to a module
